+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 ### Added
 
 - [TD-4045] Rule implementation `execute` action
-- [TD-4230] GraphQL `Templates` schema preprocesses templates
+- [TD-4548] Upload/Download csv notes with multiple selection fields separted by `|`
 - [TD-4660] Support for reference data
+- [TD-4230] GraphQL `Templates` schema preprocesses templates
 
 ### Fixed